requirements.txt
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-openai==0.3.4
langchain-tests==0.3.11
langchain-text-splitters==0.3.6
langchainhub==0.1.21
langsmith==0.3.8
langchain-openai
langchain-qdrant
qdrant-client
openai
streamlit
pandas
python-dotenv
tqdm
plotly
numpy